Fix suspense fetching layout
